% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_stream_server.R
\name{lc_server_iface}
\alias{lc_server_iface}
\title{livecode server interface}
\description{
This is a high level, user facing interface class that allows
for the creation of a livecode server sharing a specific file.
The interface also provides additional tools for sending messages.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{url}}{The current url of the server.}

\item{\code{path}}{The path of the file being served.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{lc_server_iface$new()}}
\item \href{#method-open}{\code{lc_server_iface$open()}}
\item \href{#method-print}{\code{lc_server_iface$print()}}
\item \href{#method-send_msg}{\code{lc_server_iface$send_msg()}}
\item \href{#method-is_running}{\code{lc_server_iface$is_running()}}
\item \href{#method-start}{\code{lc_server_iface$start()}}
\item \href{#method-stop}{\code{lc_server_iface$stop()}}
\item \href{#method-restart}{\code{lc_server_iface$restart()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new livecode server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lc_server_iface$new(
  file,
  ip,
  port,
  interval = 2,
  bitly = FALSE,
  auto_save = TRUE,
  open_browser = TRUE,
  public_url = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{Path to file to broadcast.}

\item{\code{ip}}{ip address to bind the server to, defaults to `0.0.0.0`.}

\item{\code{port}}{port of the server, defaults to a random value.}

\item{\code{interval}}{page update interval in seconds.}

\item{\code{bitly}}{should a bitly bit link be created for the server.}

\item{\code{auto_save}}{should the broadcast file be auto saved update tic.}

\item{\code{open_browser}}{should a browser session be opened.}

\item{\code{public_url}}{public URL to advertise to viewers, e.g. an ngrok URL.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-open"></a>}}
\if{latex}{\out{\hypertarget{method-open}{}}}
\subsection{Method \code{open()}}{
Open server in browser
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lc_server_iface$open()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Class print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lc_server_iface$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-send_msg"></a>}}
\if{latex}{\out{\hypertarget{method-send_msg}{}}}
\subsection{Method \code{send_msg()}}{
Send a noty message to all connected users on the next update tic.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lc_server_iface$send_msg(
  text,
  type = "info",
  theme = "bootstrap-v4",
  layout = "topRight",
  ...,
  parse_md = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{text of the message.}

\item{\code{type}}{message type (`alert`, `success`, `warning`, `error`, `info`).}

\item{\code{theme}}{message theme (See [here](https://ned.im/noty/#/themes) for options)}

\item{\code{layout}}{message location.}

\item{\code{...}}{additional noty arguments.}

\item{\code{parse_md}}{should message text be processed as markdown before sending.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_running"></a>}}
\if{latex}{\out{\hypertarget{method-is_running}{}}}
\subsection{Method \code{is_running()}}{
Determine if the server is running.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lc_server_iface$is_running()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns `TRUE` if the server is running.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start"></a>}}
\if{latex}{\out{\hypertarget{method-start}{}}}
\subsection{Method \code{start()}}{
Start the server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lc_server_iface$start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\if{latex}{\out{\hypertarget{method-stop}{}}}
\subsection{Method \code{stop()}}{
Stop the server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lc_server_iface$stop(warn = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{warn}}{Should the users be sent a warning that the server is shutting down.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-restart"></a>}}
\if{latex}{\out{\hypertarget{method-restart}{}}}
\subsection{Method \code{restart()}}{
Restart the server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lc_server_iface$restart()}\if{html}{\out{</div>}}
}

}
}
