% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers.R
\name{list_servers}
\alias{list_servers}
\title{List livecodeR servers created in this R session}
\usage{
list_servers()
}
\value{
A list of server objects.
}
\description{
List livecodeR servers created in this R session
}
