% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_stream_server.R
\name{LiveCodeServer}
\alias{LiveCodeServer}
\title{A local source-file streaming server}
\description{
A local source-file streaming server

A local source-file streaming server
}
\details{
This class is returned by [serve_file()]. Use `$stop()`, `$start()`, or
`$restart()` to manage the server.
}
\keyword{internal}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{url}}{Local URL that can be opened on the presenting computer.}

\item{\code{path}}{Absolute path to the streamed file.}

\item{\code{host}}{Listening address.}

\item{\code{port}}{Listening port.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LiveCodeServer-new}{\code{LiveCodeServer$new()}}
\item \href{#method-LiveCodeServer-start}{\code{LiveCodeServer$start()}}
\item \href{#method-LiveCodeServer-stop}{\code{LiveCodeServer$stop()}}
\item \href{#method-LiveCodeServer-restart}{\code{LiveCodeServer$restart()}}
\item \href{#method-LiveCodeServer-is_running}{\code{LiveCodeServer$is_running()}}
\item \href{#method-LiveCodeServer-print}{\code{LiveCodeServer$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LiveCodeServer-new"></a>}}
\if{latex}{\out{\hypertarget{method-LiveCodeServer-new}{}}}
\subsection{Method \code{new()}}{
Create and start a streaming server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LiveCodeServer$new(
  path,
  host,
  port,
  interval,
  open_browser,
  document_id = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Normalized path to the streamed file.}

\item{\code{host}}{Address on which the server listens.}

\item{\code{port}}{TCP port used by the server.}

\item{\code{interval}}{Browser polling interval in seconds.}

\item{\code{open_browser}}{Whether to open the local viewer.}

\item{\code{document_id}}{Optional RStudio document identifier to auto-save.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LiveCodeServer-start"></a>}}
\if{latex}{\out{\hypertarget{method-LiveCodeServer-start}{}}}
\subsection{Method \code{start()}}{
Start the server if it is stopped.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LiveCodeServer$start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LiveCodeServer-stop"></a>}}
\if{latex}{\out{\hypertarget{method-LiveCodeServer-stop}{}}}
\subsection{Method \code{stop()}}{
Stop the server immediately.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LiveCodeServer$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LiveCodeServer-restart"></a>}}
\if{latex}{\out{\hypertarget{method-LiveCodeServer-restart}{}}}
\subsection{Method \code{restart()}}{
Stop and then start the server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LiveCodeServer$restart()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LiveCodeServer-is_running"></a>}}
\if{latex}{\out{\hypertarget{method-LiveCodeServer-is_running}{}}}
\subsection{Method \code{is_running()}}{
Report whether the underlying HTTP server is running.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LiveCodeServer$is_running()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LiveCodeServer-print"></a>}}
\if{latex}{\out{\hypertarget{method-LiveCodeServer-print}{}}}
\subsection{Method \code{print()}}{
Print the server status, file, and local URL.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LiveCodeServer$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Unused.}
}
\if{html}{\out{</div>}}
}
}
}
