% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livecodeR.R
\docType{package}
\name{livecodeR-package}
\alias{livecodeR}
\alias{livecodeR-package}
\title{Stream source code from a local web server}
\description{
The livecodeR package serves one local source file to one or more browsers.
It does not create public tunnels; tools such as ngrok can expose the local
HTTP server separately.
}
\author{
\strong{Maintainer}: Colin Rundel \email{rundel@gmail.com}

}
\keyword{internal}
