% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_stream_server.R
\name{serve_file}
\alias{serve_file}
\title{Create a livecode server for broadcasting a file}
\usage{
serve_file(
  file,
  ip,
  port,
  interval = 1,
  bitly = FALSE,
  auto_save = TRUE,
  open_browser = TRUE,
  public_url = NULL
)
}
\arguments{
\item{file}{Path to file to broadcast.}

\item{ip}{ip address to bind the server to, defaults to `0.0.0.0`.}

\item{port}{port of the server, defaults to a random value.}

\item{interval}{page update interval in seconds.}

\item{bitly}{should a bitly bit link be created for the server.}

\item{auto_save}{should the broadcast file be auto saved during each update tic.}

\item{open_browser}{should a browser session be opened.}

\item{public_url}{public URL to advertise to viewers, e.g. an ngrok URL.}
}
\description{
Create a livecode server for broadcasting a file
}
