% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers.R
\name{stop_all}
\alias{stop_all}
\title{Stop every livecodeR server created in this R session}
\usage{
stop_all()
}
\value{
Invisibly returns `NULL`.
}
\description{
Stop every livecodeR server created in this R session
}
